refactor(core): move user model to /core/model
Refactor in folder structure to follow the guide

Closes #8
