+++ src/app/app.component.ts
@@ -2,7 +2,7 @@ import { Component, OnInit, OnDestroy } from '@angular/core';
 import { AngularFireAuth } from 'angularfire2/auth';
 import { Router } from '@angular/router';
 import { Subscription } from 'rxjs/Subscription';
-import { User } from './shared/login/user.model';
+import { User } from './core/model/user.model';
 
 @Component({
 	selector: 'app-root',
