+++ src/app/core/model/user.model.ts
@@ -1,0 +1,8 @@
+export class User {
+	constructor(
+		public uid?: string,
+		public Name?: string,
+		public Email?: string,
+		public Password?: string
+	) { }
+}
