+++ src/app/gains/gain-list/gain-list.component.ts
@@ -6,7 +6,7 @@ import { AngularFireDatabase } from 'angularfire2/database';
 
 import { toast } from 'angular2-materialize';
 import { Gain } from '../shared/gain.model';
-import { User } from '../../shared/login/user.model';
+import { User } from '../../core/model/user.model';
 
 @Component({
 	selector: 'app-gain-list',
