--- src/app/shared/login/user.model.ts
@@ -1,8 +1,0 @@
-export class User {
-    constructor(
-        public uid : string = "",
-        public Name: string = "",
-        public Email: string = "",
-        public Password: string = ""
-    ) { }
-}
